+++ HISTORY.rst
@@ -6,7 +6,7 @@ unreleased
 ------------------
 
 * Improve filterset class initialization during fetching export queryset
-* Make resource first filter, then order queryset
+* Add ability to customize the order of operation for export queryset
 * Add ability to set dataset title
 
 1.8.0 (2025-06-30)